pipeline/requirements.txt
alembic==1.16.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.3
celery-types==0.23.0
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
greenlet==3.2.3
grpcio==1.73.1
grpcio-tools==1.73.1
idna==3.10
iniconfig==2.1.0
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
paramiko==3.5.1
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyNaCl==1.5.0
pysftp==0.2.9
pytest==8.4.1
pytest-dependency==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.2.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.41.2
starlette==0.47.1
types-grpcio==1.0.0.20250703
types-paramiko==3.5.0.20250516
types-protobuf==6.30.2.20250703
types-pysftp==0.2.17.20241117
types-sqlalchemy-utils==1.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
